refactor request tracing
