--- src/routes/request.rs
@@ -1,366 +1,0 @@
-use aide::axum::{
-    routing::{head, post, put},
-    ApiRouter,
-};
-use axum::{
-    extract::Path,
-    http::{HeaderMap, Method, StatusCode},
-    Extension,
-};
-use axum_jsonschema::Json;
-use redis::{aio::ConnectionManager, AsyncCommands, ExistenceCheck, SetExpiry, SetOptions};
-use schemars::JsonSchema;
-use std::env;
-use std::str::FromStr;
-use std::sync::Arc;
-use tower_http::cors::{AllowHeaders, Any, CorsLayer};
-use uuid::Uuid;
-
-use crate::observability;
-use crate::utils::{
-    handle_redis_error, validate_request_id, AppOverrides, RequestPayload, RequestStatus,
-    EXPIRE_AFTER_SECONDS, REQ_STATUS_PREFIX,
-};
-
-const REQ_PREFIX: &str = "req:";
-const ACCEPT_IDKIT_FLOW_ID_HEADER: &str = "accept-idkit-flow-id";
-
-#[derive(Debug, serde::Deserialize, JsonSchema)]
-struct CreateRequestBody {
-    /// The initialization vector for the encrypted payload (opaque to the bridge).
-    iv: String,
-    /// The encrypted payload (opaque to the bridge).
-    payload: String,
-    /// Optional client-supplied `request_id`. When present, the bridge stores
-    /// the request under this key with NX semantics (409 on collision); when
-    /// absent, the bridge generates a UUID v4. Lets the RP address requests by
-    /// any opaque identifier they choose (e.g. an HKDF output) so the bridge
-    /// stays a generic content-addressable single-use store rather than baking
-    /// in any specific application's flow.
-    ///
-    /// **Entropy is the RP's responsibility.** The bridge enforces length
-    /// bounds and a charset (see `validate_request_id`) but does not check
-    /// cryptographic randomness. A predictable identifier lets a third party
-    /// GETDEL the request before the legitimate consumer does — confidentiality
-    /// still holds because the payload is encrypted, but the single-use
-    /// guarantee is broken. RPs should use a high-entropy source (UUID v4,
-    /// HKDF output, etc.). TODO: enforce a stronger guarantee at the protocol
-    /// layer — e.g. a server-mixed nonce — once we have a clear story for it.
-    #[serde(default)]
-    request_id: Option<String>,
-    /// Opt-in capability flag. The bridge only surfaces `app_overrides` in the
-    /// response when the client explicitly advertises that it can parse it.
-    ///
-    /// Adding a field to the response is **not** safe for clients that reject
-    /// unknown JSON keys (e.g. a strict `kotlinx.serialization` decoder), and
-    /// those clients are already shipped — they can't be retroactively fixed.
-    /// Since `POST /request` carries no platform/version, gate the field behind
-    /// an explicit flag instead: clients that predate the feature omit it and
-    /// get the lean `{request_id}` response, so a server-side override rollout
-    /// can never break them. Updated SDKs send `true` to receive overrides.
-    #[serde(default)]
-    supports_app_overrides: bool,
-}
-
-#[derive(Debug, serde::Serialize, JsonSchema)]
-struct RequestCreatedPayload {
-    /// The unique identifier for the request — the client-supplied value if
-    /// one was provided, otherwise a server-generated UUID v4.
-    request_id: String,
-    /// Temporary workaround for the World ID app rollout: lets the server
-    /// configure deeplink values and app-specific overrides. Only populated for
-    /// clients that opt in via `supports_app_overrides`; omitted entirely
-    /// otherwise (empty map) so unaware clients see the legacy `{request_id}`
-    /// response shape.
-    #[serde(skip_serializing_if = "AppOverrides::is_empty")]
-    app_overrides: AppOverrides,
-}
-
-#[derive(Debug, serde::Serialize, JsonSchema)]
-struct RequestResponse {
-    /// The initialization vector for the encrypted payload.
-    iv: String,
-    /// The encrypted payload.
-    payload: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    idkit_flow_id: Option<Uuid>,
-}
-
-pub fn handler() -> ApiRouter {
-    let cors = CorsLayer::new()
-        .allow_origin(Any)
-        .allow_headers(AllowHeaders::any())
-        .allow_methods([Method::GET, Method::POST, Method::HEAD, Method::PUT]);
-
-    let environment = env::var("ENVIRONMENT")
-        .unwrap_or_else(|_| "unknown".to_string())
-        .trim()
-        .to_lowercase();
-
-    // Base routes
-    let mut router = ApiRouter::new()
-        .api_route("/request", post(insert_request_handler))
-        .api_route(
-            "/request/:request_id",
-            head(has_request).get(get_request_handler),
-        )
-        .layer(cors);
-
-    // Only enable PUT in staging
-    if environment == "staging" {
-        router = router.api_route("/request/:request_id", put(put_request));
-    }
-
-    router
-}
-
-async fn has_request(
-    Path(request_id): Path<String>,
-    Extension(mut redis): Extension<ConnectionManager>,
-) -> StatusCode {
-    let request_id = request_id.to_lowercase();
-    if validate_request_id(&request_id).is_err() {
-        return StatusCode::BAD_REQUEST;
-    }
-
-    let Ok(exists) = redis
-        .exists::<_, bool>(format!("{REQ_PREFIX}{request_id}"))
-        .await
-    else {
-        return StatusCode::INTERNAL_SERVER_ERROR;
-    };
-
-    if exists {
-        StatusCode::OK
-    } else {
-        StatusCode::NOT_FOUND
-    }
-}
-
-async fn get_request_handler(
-    Path(request_id): Path<String>,
-    Extension(mut redis): Extension<ConnectionManager>,
-    headers: HeaderMap,
-) -> Result<Json<RequestResponse>, StatusCode> {
-    let request_id = request_id.to_lowercase();
-    let accept_idkit_flow_id = headers.contains_key(ACCEPT_IDKIT_FLOW_ID_HEADER);
-    let (payload, current_status) =
-        get_request(&mut redis, &request_id, accept_idkit_flow_id).await?;
-
-    // Keep request identifiers in the pre-existing operational logs outside
-    // the bounded root span created by `get_request`.
-    tracing::info!(
-        "Request {request_id} state transition: {} -> {}",
-        current_status,
-        RequestStatus::Retrieved
-    );
-
-    Ok(payload)
-}
-
-#[tracing::instrument(
-    parent = None,
-    name = "wallet_bridge.request.consume",
-    skip_all,
-    fields(
-        idkit_flow_id = tracing::field::Empty,
-        http.route = "/request/:request_id",
-        queue_age_ms = tracing::field::Empty,
-    )
-)]
-async fn get_request(
-    redis: &mut ConnectionManager,
-    request_id: &str,
-    accept_idkit_flow_id: bool,
-) -> Result<(Json<RequestResponse>, RequestStatus), StatusCode> {
-    validate_request_id(request_id)?;
-
-    // MULTI/EXEC keeps the status, one-time payload consumption, and flow
-    // metadata snapshot consistent without requiring Redis Lua.
-    let mut pipe = redis::pipe();
-    pipe.atomic()
-        .get(format!("{REQ_STATUS_PREFIX}{request_id}"))
-        .get_del(format!("{REQ_PREFIX}{request_id}"))
-        .get(observability::flow_key(request_id));
-
-    let (status, value, flow): (Option<String>, Option<Vec<u8>>, Option<String>) =
-        pipe.query_async(redis).await.map_err(handle_redis_error)?;
-
-    let current_status = status
-        .and_then(|status| RequestStatus::from_str(&status).ok())
-        .unwrap_or(RequestStatus::Initialized);
-
-    let value = value.ok_or(StatusCode::NOT_FOUND)?;
-    let RequestPayload { iv, payload } =
-        serde_json::from_slice(&value).map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
-
-    // Preserve the existing status behavior: consuming the payload resets
-    // the status TTL and a failure remains a request failure.
-    redis
-        .set_ex::<_, _, ()>(
-            format!("{REQ_STATUS_PREFIX}{request_id}"),
-            RequestStatus::Retrieved.to_string(),
-            EXPIRE_AFTER_SECONDS,
-        )
-        .await
-        .map_err(handle_redis_error)?;
-
-    let idkit_flow_id =
-        observability::observe_request_handoff(redis, request_id, flow.as_deref()).await;
-
-    Ok((
-        Json(RequestResponse {
-            iv,
-            payload,
-            idkit_flow_id: if accept_idkit_flow_id {
-                idkit_flow_id
-            } else {
-                None
-            },
-        }),
-        current_status,
-    ))
-}
-
-/// Create a new request. Optionally accepts a client-supplied `request_id`
-/// with NX semantics; otherwise generates a UUID v4.
-async fn insert_request_handler(
-    Extension(mut redis): Extension<ConnectionManager>,
-    Extension(app_overrides): Extension<Arc<AppOverrides>>,
-    Json(body): Json<CreateRequestBody>,
-) -> Result<Json<RequestCreatedPayload>, StatusCode> {
-    let payload = insert_request(&mut redis, &app_overrides, body).await?;
-
-    // Keep request identifiers in the pre-existing operational logs, outside
-    // the bounded root span created by `insert_request`.
-    tracing::info!("Processing /request: {}", payload.request_id);
-    tracing::info!(
-        "Request {} state transition: new -> {}",
-        payload.request_id,
-        RequestStatus::Initialized
-    );
-    tracing::info!("Successfully processed /request: {}", payload.request_id);
-
-    Ok(Json(payload))
-}
-
-#[tracing::instrument(
-    parent = None,
-    name = "wallet_bridge.request.create",
-    skip_all,
-    fields(
-        idkit_flow_id = tracing::field::Empty,
-        http.route = "/request",
-    )
-)]
-async fn insert_request(
-    redis: &mut ConnectionManager,
-    app_overrides: &AppOverrides,
-    body: CreateRequestBody,
-) -> Result<RequestCreatedPayload, StatusCode> {
-    let CreateRequestBody {
-        iv,
-        payload,
-        request_id,
-        supports_app_overrides,
-    } = body;
-
-    let request_id = match request_id {
-        Some(id) => {
-            let id = id.to_lowercase();
-            validate_request_id(&id)?;
-            id
-        }
-        None => Uuid::new_v4().to_string(),
-    };
-
-    let payload = RequestPayload::new(iv, payload);
-    let payload_bytes =
-        serde_json::to_vec(&payload).map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?;
-
-    // SET NX on the payload — collisions return 409 in a single round trip.
-    let options = SetOptions::default()
-        .conditional_set(ExistenceCheck::NX)
-        .with_expiration(SetExpiry::EX(EXPIRE_AFTER_SECONDS));
-
-    let set_ok: Option<String> = redis
-        .set_options(format!("{REQ_PREFIX}{request_id}"), payload_bytes, options)
-        .await
-        .map_err(handle_redis_error)?;
-
-    if set_ok.is_none() {
-        return Err(StatusCode::CONFLICT);
-    }
-
-    redis
-        .set_ex::<_, _, ()>(
-            format!("{REQ_STATUS_PREFIX}{request_id}"),
-            RequestStatus::Initialized.to_string(),
-            EXPIRE_AFTER_SECONDS,
-        )
-        .await
-        .map_err(handle_redis_error)?;
-
-    if let Err(outcome) = observability::mint_and_store_idkit_flow_id(redis, &request_id).await {
-        tracing::warn!(outcome, "Failed to mint and store IDKit flow ID");
-    }
-
-    Ok(RequestCreatedPayload {
-        request_id,
-        app_overrides: select_response_overrides(supports_app_overrides, app_overrides),
-    })
-}
-
-/// Pick the overrides to echo back on `POST /request`: the configured map for
-/// clients that opted in via `supports_app_overrides`, otherwise an empty map
-/// (omitted from the response). Keeps the field invisible to clients that
-/// predate the feature so a server-side override rollout can't break a strict,
-/// unknown-key-rejecting JSON parser.
-fn select_response_overrides(opted_in: bool, configured: &AppOverrides) -> AppOverrides {
-    if opted_in {
-        configured.clone()
-    } else {
-        AppOverrides::default()
-    }
-}
-
-/// Create a new request by ID idempotently — retries succeed, even if the request exists.
-/// Note: only enabled in staging.
-async fn put_request(
-    Path(request_id): Path<String>,
-    Extension(mut redis): Extension<ConnectionManager>,
-    Json(request): Json<RequestPayload>,
-) -> Result<StatusCode, StatusCode> {
-    let request_id = request_id.to_lowercase();
-    validate_request_id(&request_id)?;
-
-    tracing::info!("Processing PUT /request: {request_id}");
-
-    // Same logic as POST, but always overwrites the existing payload, sets status, and resets the TTL.
-    redis
-        .set_ex::<_, _, ()>(
-            format!("{REQ_STATUS_PREFIX}{request_id}"),
-            RequestStatus::Initialized.to_string(),
-            EXPIRE_AFTER_SECONDS,
-        )
-        .await
-        .map_err(handle_redis_error)?;
-
-    tracing::info!(
-        "Request {request_id} state transition: new -> {}",
-        RequestStatus::Initialized
-    );
-
-    redis
-        .set_ex::<_, _, ()>(
-            format!("{REQ_PREFIX}{request_id}"),
-            serde_json::to_vec(&request).map_err(|_| StatusCode::INTERNAL_SERVER_ERROR)?,
-            EXPIRE_AFTER_SECONDS,
-        )
-        .await
-        .map_err(handle_redis_error)?;
-
-    tracing::info!("Successfully PUT /request: {request_id}");
-
-    Ok(StatusCode::CREATED)
-}
