+++ tests/common.rs
@@ -53,20 +53,11 @@ async fn send(
     method: Method,
     route: &str,
     body: Option<&Value>,
-) -> (u16, String) {
-    send_with_headers(app, method, route, body, &[]).await
-}
-
-async fn send_with_headers(
-    app: &axum::Router,
-    method: Method,
-    route: &str,
-    body: Option<&Value>,
-    headers: &[(&str, &str)],
+    header: Option<(&str, &str)>,
 ) -> (u16, String) {
     let mut builder = Request::builder().uri(route).method(method);
-    for (name, value) in headers {
-        builder = builder.header(*name, *value);
+    if let Some((name, value)) = header {
+        builder = builder.header(name, value);
     }
     let request = match body {
         Some(json) => builder
@@ -95,7 +86,7 @@ async fn send_with_headers(
 }
 
 pub async fn get(app: &axum::Router, route: &str) -> (u16, String) {
-    send(app, Method::GET, route, None).await
+    send(app, Method::GET, route, None, None).await
 }
 
 pub async fn get_with_header(
@@ -104,13 +95,13 @@ pub async fn get_with_header(
     name: &str,
     value: &str,
 ) -> (u16, String) {
-    send_with_headers(app, Method::GET, route, None, &[(name, value)]).await
+    send(app, Method::GET, route, None, Some((name, value))).await
 }
 
 pub async fn post(app: &axum::Router, route: &str, body: &Value) -> (u16, String) {
-    send(app, Method::POST, route, Some(body)).await
+    send(app, Method::POST, route, Some(body), None).await
 }
 
 pub async fn put(app: &axum::Router, route: &str, body: &Value) -> (u16, String) {
-    send(app, Method::PUT, route, Some(body)).await
+    send(app, Method::PUT, route, Some(body), None).await
 }
